+++ examples/plot_speed_traces.py
@@ -48,6 +48,6 @@ ax.set_ylabel('Speed in km/h')
 
 ax.legend()
 plt.suptitle(f"Fastest Lap Comparison \n "
-             f"{session.event.name} {session.event.year} Qualifying")
+             f"{session.event['EventName']} {session.event.year} Qualifying")
 
 plt.show()